pyproject.toml
[project]
name = "AnimaGo"
version = "0.1.0"
description = "Gotta photograph em all"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Shresht Bhowmick", email = "shresht@mit.edu" },
    { name = "Tyler Dong", email = "tylerdong878@gmail.com" },
    { name = "Raymond Xu", email = "Ray01778@gmail.com" }
]
dependencies = [
    "fastapi>=0.115.8",
    "firebase-admin>=6.6.0",
    "flet[all]>=0.27.0",
    "folium>=0.19.4",
    "geopy>=2.4.1",
    "moondream>=0.0.6",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pillow>=10.4.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "segment-anything",
    "sounddevice>=0.5.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "ultralytics==8.0.0",
    "websockets>=15.0",
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "tetraslam.com"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "AnimaGo"

# company name to display in about app dialogs
company = "Tetracorp, Inc."

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by Tetracorp, Inc."

[tool.flet.app]
path = "src"

[tool.uv]
dev-dependencies = [
    "flet[all]==0.27.0",
]

[tool.uv.sources]
segment-anything = { git = "https://github.com/facebookresearch/segment-anything.git" }

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.27.0"}
